% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMat.R
\name{aperm.tables}
\alias{aperm.tables}
\title{Permute dimensions of tables}
\usage{
\method{aperm}{tables}(a, perm)
}
\arguments{
\item{a}{object of class \code{tables}}

\item{perm}{permutation of 1,...,k, where each table has k dimensions}
}
\description{
Method for permuting indices of tables object.
}

