% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as.matrix.tables}
\alias{as.matrix.tables}
\title{Convert tables into matrix}
\usage{
\method{as.matrix}{tables}(x, ...)
}
\arguments{
\item{x}{\code{tables} object}

\item{...}{other arguments}
}
\value{
A \code{matrix} object
}
\description{
Convert tables into matrix
}
