% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_tables}
\alias{as_tables}
\title{As tables}
\usage{
as_tables(x, tdim, conditional, ...)
}
\arguments{
\item{x}{array or matrix object}

\item{tdim}{dimensions for each table}

\item{conditional}{integer vector of indices that are conditional}

\item{...}{other arguments for methods}
}
\value{
A \code{tables} object.
}
\description{
As tables
}
