% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capply.R
\name{capply}
\alias{capply}
\title{Apply function over tables}
\usage{
capply(x, f, ...)
}
\arguments{
\item{x}{object of class \code{tables}}

\item{f}{function to apply to each table}

\item{...}{additional arguments to \code{f}}
}
\description{
Apply function over tables
}

