% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{entropy}
\alias{entropy}
\alias{entropy.default}
\alias{entropy.array}
\alias{entropy.tables}
\title{Calculate entropy of discrete distribution}
\usage{
entropy(p, ...)

\method{entropy}{default}(p, ...)

\method{entropy}{array}(p, margin, ...)

\method{entropy}{tables}(p, margin, ...)
}
\arguments{
\item{p}{non-negative numeric vector}

\item{...}{other arguments to methods}

\item{margin}{margin to consider}
}
\value{
A numeric value of the entopy, or vector of entropies.
}
\description{
Calculate entropy of discrete distribution
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for vectors

\item \code{array}: Method for arrays

\item \code{tables}: Method for \code{tables} object
}}

