% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rev_tables.R
\name{is_rev}
\alias{is_rev}
\title{Test if tables object is reversed}
\usage{
is_rev(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
Checks if the rows of the underlying matrix represent each table or each entry in a table
}
