% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{is_tables}
\alias{is_tables}
\title{Test object is a collection of tables}
\usage{
is_tables(x)
}
\arguments{
\item{x}{object to be tested}
}
\description{
Checks if \code{"tables"} is in the list of class attributes.
}
