% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{kl}
\alias{kl}
\alias{kl.default}
\alias{kl.tables}
\title{Kullback-Leibler Divergence}
\usage{
kl(x, y, ...)

\method{kl}{default}(x, y, ...)

\method{kl}{tables}(x, y, ...)
}
\arguments{
\item{x, y}{vectors (of probabilities)}

\item{...}{other arguments to methods}
}
\value{
a numberic value, vector or matrix of KL-divergences.
}
\description{
Get the KL Divergence between two discrete distributions
}
\section{Methods (by class)}{
\itemize{
\item \code{kl(default)}: Default method for vectors

\item \code{kl(tables)}: Method for \code{tables} object

}}
