% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin}
\alias{margin}
\alias{margin2}
\alias{conditional}
\alias{conditional2}
\alias{intervention}
\title{Get margin of a table or tables}
\usage{
margin(x, ...)

margin2(x, ...)

conditional(x, ...)

conditional2(x, ...)

intervention(x, ...)
}
\arguments{
\item{x}{a contingency table or \code{tables} object}

\item{...}{a contingency table or \code{tables} object}
}
\description{
Get margin of a table or tables
}
\details{
\code{margin2} keeps all dimensions, and 
hence results will sum to the number of cells summed over.
}
\section{Functions}{
\itemize{
\item \code{margin2}: keep all dimensions

\item \code{conditional}: conditional distributions

\item \code{conditional2}: conditional distributions with all dimensions kept

\item \code{intervention}: interventional distributions
}}

