% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{multiInf}
\alias{multiInf}
\alias{multiInf.default}
\alias{multiInf.tables}
\title{Multiinformation}
\usage{
multiInf(x, ...)

\method{multiInf}{default}(x, margin = NULL, ...)

\method{multiInf}{tables}(x, margin = NULL, ...)
}
\arguments{
\item{x}{vectors (of probabilities)}

\item{...}{other arguments to methods}

\item{margin}{margin to find multiinformation for}
}
\value{
a numberic value, vector or matrix of required multiinformation.
}
\description{
Get the multiinformation for a discrete distribution
}
\section{Methods (by class)}{
\itemize{
\item \code{multiInf(default)}: Default method for vectors and arrays

\item \code{multiInf(tables)}: Method for \code{tables} object

}}
