% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{mutualInf}
\alias{mutualInf}
\alias{mutualInf.default}
\alias{mutualInf.tables}
\title{(Conditional) mutual information}
\usage{
mutualInf(p, m1, m2, condition, ...)

\method{mutualInf}{default}(p, m1, m2, condition, ...)

\method{mutualInf}{tables}(p, m1, m2, condition, ...)
}
\arguments{
\item{p}{numeric array or \code{tables} class}

\item{m1, m2}{margins for mutual information}

\item{condition}{conditional margin}

\item{...}{other arguments to methods}
}
\description{
(Conditional) mutual information
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for vectors

\item \code{tables}: Method for \code{tables} object
}}

