% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{ntables}
\alias{ntables}
\title{Number of tables}
\usage{
ntables(x)
}
\arguments{
\item{x}{an object of class \code{tables}}
}
\value{
An integer.
}
\description{
Number of tables
}
\details{
Gives the number of tables in an object of class
\code{tables}.
}
