% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{print.tables}
\alias{print.tables}
\title{Print tables}
\usage{
\method{print}{tables}(x, ...)
}
\arguments{
\item{x}{object of class \code{tables}}

\item{...}{arguments to pass to print method for an array}
}
\value{
The input provided (invisibly).
}
\description{
Print method for object of class \code{tables}.
}
