% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphs.R
\name{rDAGmodel}
\alias{rDAGmodel}
\title{Generate probability distributions according to a graphical model}
\usage{
rDAGmodel(graph, n, dim = 2L, alpha = 1)
}
\arguments{
\item{graph}{object of class \code{mixedgraph} giving a directed acyclic graph}

\item{n}{number of distributions to generate}

\item{dim}{integer vector of dimensions for the distribution}

\item{alpha}{parameter controlling Dirichlet distributions}
}
\description{
Generate tables which factorize according to a directed acyclic graph.
}
\examples{
library(MixedGraphs)
gr <- mixedgraph(5, edges=list(directed=list(c(1,2),c(1,3),c(2,4),c(3,5),c(4,5))))
rDAGmodel(gr, 1)
}

