% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capply.R
\name{repTables}
\alias{repTables}
\title{Turn distributions into tables}
\usage{
repTables(n, f, ..., rev = FALSE)
}
\arguments{
\item{n}{number of distributions to generate}

\item{f}{function that generates a probability distribution}

\item{...}{arguments to \code{f}}

\item{rev}{logical: should output move through each table fastest?}
}
\value{
a tables object containing the outputs of \code{f}
}
\description{
Turn distributions into tables
}
