% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{tables}
\alias{tables}
\title{Create blank tables}
\usage{
tables(n, tdim, rev = FALSE)
}
\arguments{
\item{n}{number of tables}

\item{tdim}{dimension of each table}

\item{rev}{logical: should output move through each table fastest?}
}
\description{
Create blank tables
}
