% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/probMat.R
\name{tdim}
\alias{tdim}
\title{Dimension of distributions over contingency tables}
\usage{
tdim(x)
}
\arguments{
\item{x}{an object of class \code{tables}}
}
\value{
an integer vector of the dimensions
}
\description{
Dimension of distributions over contingency tables
}
\details{
The class \code{tables} is used to represent a collection of 
multidimentional tables; this function
returns the dimension of each table.
}

