% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as.array.tables}
\alias{as.array.tables}
\title{Convert tables into array}
\usage{
\method{as.array}{tables}(x, ...)
}
\arguments{
\item{x}{\code{tables} object}

\item{...}{other arguments}
}
\value{
An \code{array} object
}
\description{
Convert tables into array
}
