% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tables.R
\name{as_tables}
\alias{as_tables}
\title{As tables}
\usage{
as_tables(x, tdim, conditional, rev = FALSE, ...)
}
\arguments{
\item{x}{array or matrix object}

\item{tdim}{dimensions for each table}

\item{conditional}{integer vector of indices that are conditional}

\item{rev}{logical: should output move through each table fastest?}

\item{...}{other arguments for methods}
}
\value{
A \code{tables} object.
}
\description{
As tables
}
\details{
Transforms a vector, matrix or array into a tables object with
the specified dimensions.  Note that if \code{rev = TRUE} then the final dimension
is (by default) used to index the tables objects, and otherwise the first
dimension..
}
