% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{interactionInf}
\alias{interactionInf}
\alias{interactionInf.default}
\title{Interaction information}
\usage{
interactionInf(p, ...)

\method{interactionInf}{default}(p, ..., condition)
}
\arguments{
\item{p}{object to find interaction information for}

\item{...}{other arguments to methods}

\item{condition}{variables on which to condition}
}
\value{
Numeric value for interaction information, or a vector of
interaction information values.
}
\description{
Interaction information
}
\section{Methods (by class)}{
\itemize{
\item \code{interactionInf(default)}: Default method for vectors

}}
