% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entropy.R
\name{kl}
\alias{kl}
\alias{kl.default}
\alias{kl.tables}
\title{Kullback-Leibler Divergence}
\usage{
kl(x, y, ...)

\method{kl}{default}(x, y)

\method{kl}{tables}(x, y)
}
\arguments{
\item{x, y}{vectors (of probabilities)}
}
\description{
Get the KL Divergence between two discrete distributions
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for vectors

\item \code{tables}: Method for \code{tables} object
}}

