% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/margin.R
\name{margin.tables}
\alias{margin.tables}
\title{Get the marginal distributions}
\usage{
\method{margin}{tables}(x, margin = NULL, order = TRUE)
}
\arguments{
\item{x}{an object of class \code{probMat}}

\item{margin}{integer vector giving margin to be calculated (1 for rows, etc.)}

\item{order}{logical indicating whether resulting indices
should be in the same order as stated in \code{margin}}

\item{sorted}{logical indicating whether vector \code{margin} 
is already in numerical order}
}
\description{
Get the marginal distributions
}
\details{
Calculates marginal distributions for each entry in a \code{probMat}.
}

