% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm_dim.R
\name{perm_dim}
\alias{perm_dim}
\title{Permute indices for variable \code{k}}
\usage{
perm_dim(x, k, perm, ...)
}
\arguments{
\item{x}{array or related object}

\item{k}{index to permute}

\item{perm}{permutation to perform}

\item{...}{other arguments (not currently used)}
}
\value{
A permuted \code{array} or \code{tables} object.
}
\description{
Currently only works for binary dimensions.
}
\details{
Permutes the levels of one variable
according to the permutation given in \code{perm}.
Can be applied to matrices, arrays or tables.
}
